+++ library/core/src/test/java/com/google/android/exoplayer2/upstream/DefaultBandwidthMeterTest.java
@@ -24,7 +24,6 @@ import android.net.NetworkInfo;
 import android.net.NetworkInfo.DetailedState;
 import android.telephony.TelephonyManager;
 import com.google.android.exoplayer2.C;
-import com.google.android.exoplayer2.util.ClosedSource;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -33,10 +32,9 @@ import org.robolectric.RuntimeEnvironment;
 import org.robolectric.Shadows;
 import org.robolectric.shadows.ShadowNetworkInfo;
 
-/** Unit test for {@link CountryAndNetworkTypeBandwidthMeter}. */
-@ClosedSource(reason = "Not ready yet")
+/** Unit test for {@link DefaultBandwidthMeter}. */
 @RunWith(RobolectricTestRunner.class)
-public final class CountryAndNetworkTypeBandwidthMeterTest {
+public final class DefaultBandwidthMeterTest {
 
   private static final String FAST_COUNTRY_ISO = "EE";
   private static final String SLOW_COUNTRY_ISO = "PG";
@@ -106,13 +104,13 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void defaultInitialBitrateEstimate_forWifi_isGreaterThanEstimateFor2G() {
     setActiveNetworkInfo(networkInfoWifi);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterWifi =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeterWifi =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimateWifi = bandwidthMeterWifi.getBitrateEstimate();
 
     setActiveNetworkInfo(networkInfo2g);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter2g =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeter2g =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimate2g = bandwidthMeter2g.getBitrateEstimate();
 
     assertThat(initialEstimateWifi).isGreaterThan(initialEstimate2g);
@@ -121,13 +119,13 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void defaultInitialBitrateEstimate_forWifi_isGreaterThanEstimateFor3G() {
     setActiveNetworkInfo(networkInfoWifi);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterWifi =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeterWifi =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimateWifi = bandwidthMeterWifi.getBitrateEstimate();
 
     setActiveNetworkInfo(networkInfo3g);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter3g =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeter3g =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimate3g = bandwidthMeter3g.getBitrateEstimate();
 
     assertThat(initialEstimateWifi).isGreaterThan(initialEstimate3g);
@@ -136,13 +134,13 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void defaultInitialBitrateEstimate_forEthernet_isGreaterThanEstimateFor2G() {
     setActiveNetworkInfo(networkInfoEthernet);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterEthernet =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeterEthernet =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimateEthernet = bandwidthMeterEthernet.getBitrateEstimate();
 
     setActiveNetworkInfo(networkInfo2g);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter2g =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeter2g =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimate2g = bandwidthMeter2g.getBitrateEstimate();
 
     assertThat(initialEstimateEthernet).isGreaterThan(initialEstimate2g);
@@ -151,13 +149,13 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void defaultInitialBitrateEstimate_forEthernet_isGreaterThanEstimateFor3G() {
     setActiveNetworkInfo(networkInfoEthernet);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterEthernet =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeterEthernet =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimateEthernet = bandwidthMeterEthernet.getBitrateEstimate();
 
     setActiveNetworkInfo(networkInfo3g);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter3g =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeter3g =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimate3g = bandwidthMeter3g.getBitrateEstimate();
 
     assertThat(initialEstimateEthernet).isGreaterThan(initialEstimate3g);
@@ -166,13 +164,13 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void defaultInitialBitrateEstimate_for4G_isGreaterThanEstimateFor2G() {
     setActiveNetworkInfo(networkInfo4g);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter4g =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeter4g =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimate4g = bandwidthMeter4g.getBitrateEstimate();
 
     setActiveNetworkInfo(networkInfo2g);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter2g =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeter2g =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimate2g = bandwidthMeter2g.getBitrateEstimate();
 
     assertThat(initialEstimate4g).isGreaterThan(initialEstimate2g);
@@ -181,13 +179,13 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void defaultInitialBitrateEstimate_for4G_isGreaterThanEstimateFor3G() {
     setActiveNetworkInfo(networkInfo4g);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter4g =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeter4g =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimate4g = bandwidthMeter4g.getBitrateEstimate();
 
     setActiveNetworkInfo(networkInfo3g);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter3g =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeter3g =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimate3g = bandwidthMeter3g.getBitrateEstimate();
 
     assertThat(initialEstimate4g).isGreaterThan(initialEstimate3g);
@@ -196,13 +194,13 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void defaultInitialBitrateEstimate_for3G_isGreaterThanEstimateFor2G() {
     setActiveNetworkInfo(networkInfo3g);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter3g =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeter3g =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimate3g = bandwidthMeter3g.getBitrateEstimate();
 
     setActiveNetworkInfo(networkInfo2g);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter2g =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeter2g =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimate2g = bandwidthMeter2g.getBitrateEstimate();
 
     assertThat(initialEstimate3g).isGreaterThan(initialEstimate2g);
@@ -211,8 +209,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void defaultInitialBitrateEstimate_forOffline_isReasonable() {
     setActiveNetworkInfo(networkInfoOffline);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
 
     assertThat(initialEstimate).isGreaterThan(100_000L);
@@ -224,13 +222,13 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
       defaultInitialBitrateEstimate_forWifi_forFastCountry_isGreaterThanEstimateForSlowCountry() {
     setActiveNetworkInfo(networkInfoWifi);
     setNetworkCountryIso(FAST_COUNTRY_ISO);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterFast =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeterFast =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimateFast = bandwidthMeterFast.getBitrateEstimate();
 
     setNetworkCountryIso(SLOW_COUNTRY_ISO);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterSlow =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeterSlow =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimateSlow = bandwidthMeterSlow.getBitrateEstimate();
 
     assertThat(initialEstimateFast).isGreaterThan(initialEstimateSlow);
@@ -241,13 +239,13 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
       defaultInitialBitrateEstimate_forEthernet_forFastCountry_isGreaterThanEstimateForSlowCountry() {
     setActiveNetworkInfo(networkInfoEthernet);
     setNetworkCountryIso(FAST_COUNTRY_ISO);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterFast =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeterFast =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimateFast = bandwidthMeterFast.getBitrateEstimate();
 
     setNetworkCountryIso(SLOW_COUNTRY_ISO);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterSlow =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeterSlow =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimateSlow = bandwidthMeterSlow.getBitrateEstimate();
 
     assertThat(initialEstimateFast).isGreaterThan(initialEstimateSlow);
@@ -258,13 +256,13 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
       defaultInitialBitrateEstimate_for2G_forFastCountry_isGreaterThanEstimateForSlowCountry() {
     setActiveNetworkInfo(networkInfo2g);
     setNetworkCountryIso(FAST_COUNTRY_ISO);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterFast =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeterFast =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimateFast = bandwidthMeterFast.getBitrateEstimate();
 
     setNetworkCountryIso(SLOW_COUNTRY_ISO);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterSlow =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeterSlow =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimateSlow = bandwidthMeterSlow.getBitrateEstimate();
 
     assertThat(initialEstimateFast).isGreaterThan(initialEstimateSlow);
@@ -275,13 +273,13 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
       defaultInitialBitrateEstimate_for3G_forFastCountry_isGreaterThanEstimateForSlowCountry() {
     setActiveNetworkInfo(networkInfo3g);
     setNetworkCountryIso(FAST_COUNTRY_ISO);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterFast =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeterFast =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimateFast = bandwidthMeterFast.getBitrateEstimate();
 
     setNetworkCountryIso(SLOW_COUNTRY_ISO);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterSlow =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeterSlow =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimateSlow = bandwidthMeterSlow.getBitrateEstimate();
 
     assertThat(initialEstimateFast).isGreaterThan(initialEstimateSlow);
@@ -292,13 +290,13 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
       defaultInitialBitrateEstimate_for4g_forFastCountry_isGreaterThanEstimateForSlowCountry() {
     setActiveNetworkInfo(networkInfo4g);
     setNetworkCountryIso(FAST_COUNTRY_ISO);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterFast =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeterFast =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimateFast = bandwidthMeterFast.getBitrateEstimate();
 
     setNetworkCountryIso(SLOW_COUNTRY_ISO);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterSlow =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeterSlow =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimateSlow = bandwidthMeterSlow.getBitrateEstimate();
 
     assertThat(initialEstimateFast).isGreaterThan(initialEstimateSlow);
@@ -307,8 +305,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void initialBitrateEstimateOverwrite_whileConnectedToNetwork_setsInitialEstimate() {
     setActiveNetworkInfo(networkInfoWifi);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application)
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application)
             .setInitialBitrateEstimate(123456789)
             .build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
@@ -319,8 +317,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void initialBitrateEstimateOverwrite_whileOffline_setsInitialEstimate() {
     setActiveNetworkInfo(networkInfoOffline);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application)
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application)
             .setInitialBitrateEstimate(123456789)
             .build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
@@ -331,8 +329,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void initialBitrateEstimateOverwrite_forWifi_whileConnectedToWifi_setsInitialEstimate() {
     setActiveNetworkInfo(networkInfoWifi);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application)
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application)
             .setInitialBitrateEstimate(C.NETWORK_TYPE_WIFI, 123456789)
             .build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
@@ -344,8 +342,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   public void
       initialBitrateEstimateOverwrite_forWifi_whileConnectedToOtherNetwork_doesNotSetInitialEstimate() {
     setActiveNetworkInfo(networkInfo2g);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application)
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application)
             .setInitialBitrateEstimate(C.NETWORK_TYPE_WIFI, 123456789)
             .build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
@@ -357,8 +355,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   public void
       initialBitrateEstimateOverwrite_forEthernet_whileConnectedToEthernet_setsInitialEstimate() {
     setActiveNetworkInfo(networkInfoEthernet);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application)
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application)
             .setInitialBitrateEstimate(C.NETWORK_TYPE_ETHERNET, 123456789)
             .build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
@@ -370,8 +368,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   public void
       initialBitrateEstimateOverwrite_forEthernet_whileConnectedToOtherNetwork_doesNotSetInitialEstimate() {
     setActiveNetworkInfo(networkInfo2g);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application)
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application)
             .setInitialBitrateEstimate(C.NETWORK_TYPE_WIFI, 123456789)
             .build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
@@ -382,8 +380,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void initialBitrateEstimateOverwrite_for2G_whileConnectedTo2G_setsInitialEstimate() {
     setActiveNetworkInfo(networkInfo2g);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application)
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application)
             .setInitialBitrateEstimate(C.NETWORK_TYPE_2G, 123456789)
             .build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
@@ -395,8 +393,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   public void
       initialBitrateEstimateOverwrite_for2G_whileConnectedToOtherNetwork_doesNotSetInitialEstimate() {
     setActiveNetworkInfo(networkInfoWifi);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application)
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application)
             .setInitialBitrateEstimate(C.NETWORK_TYPE_2G, 123456789)
             .build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
@@ -407,8 +405,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void initialBitrateEstimateOverwrite_for3G_whileConnectedTo3G_setsInitialEstimate() {
     setActiveNetworkInfo(networkInfo3g);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application)
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application)
             .setInitialBitrateEstimate(C.NETWORK_TYPE_3G, 123456789)
             .build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
@@ -420,8 +418,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   public void
       initialBitrateEstimateOverwrite_for3G_whileConnectedToOtherNetwork_doesNotSetInitialEstimate() {
     setActiveNetworkInfo(networkInfoWifi);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application)
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application)
             .setInitialBitrateEstimate(C.NETWORK_TYPE_3G, 123456789)
             .build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
@@ -432,8 +430,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void initialBitrateEstimateOverwrite_for4G_whileConnectedTo4G_setsInitialEstimate() {
     setActiveNetworkInfo(networkInfo4g);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application)
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application)
             .setInitialBitrateEstimate(C.NETWORK_TYPE_4G, 123456789)
             .build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
@@ -445,8 +443,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   public void
       initialBitrateEstimateOverwrite_for4G_whileConnectedToOtherNetwork_doesNotSetInitialEstimate() {
     setActiveNetworkInfo(networkInfoWifi);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application)
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application)
             .setInitialBitrateEstimate(C.NETWORK_TYPE_4G, 123456789)
             .build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
@@ -457,8 +455,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void initialBitrateEstimateOverwrite_forOffline_whileOffline_setsInitialEstimate() {
     setActiveNetworkInfo(networkInfoOffline);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application)
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application)
             .setInitialBitrateEstimate(C.NETWORK_TYPE_OFFLINE, 123456789)
             .build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
@@ -470,8 +468,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   public void
       initialBitrateEstimateOverwrite_forOffline_whileConnectedToNetwork_doesNotSetInitialEstimate() {
     setActiveNetworkInfo(networkInfoWifi);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application)
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application)
             .setInitialBitrateEstimate(C.NETWORK_TYPE_OFFLINE, 123456789)
             .build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
@@ -482,13 +480,13 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void initialBitrateEstimateOverwrite_forCountry_usesDefaultValuesForCountry() {
     setNetworkCountryIso(SLOW_COUNTRY_ISO);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterSlow =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeterSlow =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimateSlow = bandwidthMeterSlow.getBitrateEstimate();
 
     setNetworkCountryIso(FAST_COUNTRY_ISO);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterFastWithSlowOverwrite =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application)
+    DefaultBandwidthMeter bandwidthMeterFastWithSlowOverwrite =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application)
             .setInitialBitrateEstimate(SLOW_COUNTRY_ISO)
             .build();
     long initialEstimateFastWithSlowOverwrite =
@@ -499,12 +497,11 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
 
   @Test
   public void defaultInitialBitrateEstimate_withoutContext_isReasonable() {
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterWithBuilder =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(/* context= */ null).build();
+    DefaultBandwidthMeter bandwidthMeterWithBuilder =
+        new DefaultBandwidthMeter.Builder(/* context= */ null).build();
     long initialEstimateWithBuilder = bandwidthMeterWithBuilder.getBitrateEstimate();
 
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeterWithoutBuilder =
-        new CountryAndNetworkTypeBandwidthMeter();
+    DefaultBandwidthMeter bandwidthMeterWithoutBuilder = new DefaultBandwidthMeter();
     long initialEstimateWithoutBuilder = bandwidthMeterWithoutBuilder.getBitrateEstimate();
 
     assertThat(initialEstimateWithBuilder).isGreaterThan(100_000L);
@@ -516,8 +513,8 @@ public final class CountryAndNetworkTypeBandwidthMeterTest {
   @Test
   public void defaultInitialBitrateEstimate_withoutAccessNetworkStatePermission_isReasonable() {
     Shadows.shadowOf(RuntimeEnvironment.application).denyPermissions(ACCESS_NETWORK_STATE);
-    CountryAndNetworkTypeBandwidthMeter bandwidthMeter =
-        new CountryAndNetworkTypeBandwidthMeter.Builder(RuntimeEnvironment.application).build();
+    DefaultBandwidthMeter bandwidthMeter =
+        new DefaultBandwidthMeter.Builder(RuntimeEnvironment.application).build();
     long initialEstimate = bandwidthMeter.getBitrateEstimate();
 
     assertThat(initialEstimate).isGreaterThan(100_000L);